microservices/Auth/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
passlib[bcrypt]
pyjwt
python-jose
httpx==0.25.0
eralchemy2==1.3.6
pygraphviz==1.10